api/requirements.txt
# Core Pydantic AI
pydantic-ai>=0.0.14
pydantic>=2.0.0
pydantic-settings>=2.6.1
python-dotenv>=1.0.0

# LLM Providers
anthropic>=0.39.0
openai>=1.0.0

# Database Integration
supabase>=2.10.0

# FastAPI Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Testing
pytest>=8.3.3
pytest-asyncio>=0.24.0
httpx>=0.27.0